tests/wpt/tests/css/css-tables/parsing/empty-cells-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module Test: parsing empty-cells with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-empty-cells">
<meta name="assert" content="empty-cells supports the full grammar 'show | hide'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("empty-cells", "show");
test_valid_value("empty-cells", "hide");
</script>
</body>
</html>
